Move `relay/types` to `relay/server/types` (and split out `ClientInfo` into `codec`
